+++ handler/character.go
@@ -6,7 +6,7 @@ import (
 	"net/http"
 
 	"github.com/echizenn/techtrain_CA_Tech_Dojo_Server_Side/application"
-	"github.com/echizenn/techtrain_CA_Tech_Dojo_Server_Side/repository"
+	"github.com/echizenn/techtrain_CA_Tech_Dojo_Server_Side/infrastructure"
 )
 
 func UserHoldCharacterList(w http.ResponseWriter, r *http.Request) {
@@ -21,9 +21,9 @@ func UserHoldCharacterList(w http.ResponseWriter, r *http.Request) {
 	header := r.Header
 	token := header["X-Token"][0] // なんで大文字になる？、0って明示して大丈夫？
 
-	ur := repository.NewUserRepository()
-	cr := repository.NewCharacterRepository()
-	ucr := repository.NewUsersCharactersRepository(cr)
+	ur := infrastructure.NewUserRepository()
+	cr := infrastructure.NewCharacterRepository()
+	ucr := infrastructure.NewUsersCharactersRepository(cr)
 
 	ucas := application.NewUsersCharactersApplicationService(ur, ucr)
