+++ infrastructure/characterRepository.go
@@ -1,4 +1,4 @@
-package repository
+package infrastructure
 
 import (
 	"database/sql"
